fixturi ispravil i testi pod nih toge chu-chu charlzgit add .!

+++ conftest.py
@@ -17,7 +17,6 @@ from helpers.courier_helpers import (
     build_courier_login_payload,
 )
 
-
 @pytest.fixture
 def courier_client():
     return CourierClient(BASE_URL)
@@ -26,16 +25,22 @@ def courier_client():
 @pytest.fixture
 def new_courier(courier_client):
     payload = build_unique_courier_payload()
-    response = courier_client.create_courier(payload)
-    assert response.status_code == 201
+
+    create_resp = courier_client.create_courier(payload)
+    assert create_resp.status_code == 201
+
     yield payload
+
     login_payload = build_courier_login_payload(
         payload["login"],
         payload["password"]
     )
-    login_response = courier_client.login_courier(login_payload)
-    courier_id = login_response.json()["id"]
-    courier_client.delete_courier(courier_id)
+
+    login_resp = courier_client.login_courier(login_payload)
+
+    if login_resp.status_code == 200:
+        courier_id = login_resp.json()["id"]
+        courier_client.delete_courier(courier_id)
 
 
 @pytest.fixture
@@ -44,10 +49,13 @@ def authorized_courier(courier_client, new_courier):
         new_courier["login"],
         new_courier["password"]
     )
-    response = courier_client.login_courier(login_payload)
-    assert response.status_code == 200
-    courier_id = response.json()["id"]
+
+    login_resp = courier_client.login_courier(login_payload)
+
+    assert login_resp.status_code == 200
+
     return {
         "courier": new_courier,
-        "id": courier_id
+        "login_response": login_resp,
+        "id": login_resp.json()["id"],
     }
